Refactor project to support Workspaces
