+++ atelier/examples/ob_gen_naive.rs
@@ -40,3 +40,4 @@ fn main() {
     println!("orders[1]{:?}", i_ob.asks[199].orders[1]);
     println!(" ... ");
 }
+
