+++ atelier/examples/ob_gen_synthetic.rs
@@ -58,3 +58,4 @@ fn main() {
         n_orderbooks[1].bids[0].orders[1]);
 
 }
+
