+++ atelier/src/simulation/generators.rs
